==> 00_clean_reformat.R <==
# These are the only three scripts that need to be run to make original_data into formatted_data.
# All others (02, 03, 04, 06, 07) were used for exploration and interim checking of the cleaning.


source("01_expand_level3.R")
source("05_resolve_duplicates.R")
source("08_additional_formats.R")
